Access Controlled the RescheduleBtn.js
